refactor: address PR #202 review comments
- Rename src/data/public-listing.ts → src/data/listing.ts (name was stale)
- Remove MAX_PHOTOS_PER_LISTING re-export from listing-photo-upload.server.ts
  (nobody imported it from there)
- Simplify photosForViewerRow to a ternary
- Replace custom .listing-photo-file-input CSS with global sr-only utility class
