+++ apps/www/src/routes/listing-show.css
@@ -258,16 +258,7 @@
 			color: var(--color-foreground);
 		}
 
-		& .listing-photo-file-input {
-			position: absolute;
-			inline-size: 1px;
-			block-size: 1px;
-			overflow: hidden;
-			clip: rect(0 0 0 0);
-			white-space: nowrap;
-		}
-
-		& .listing-photo-ghost--disabled {
+& .listing-photo-ghost--disabled {
 			cursor: not-allowed;
 			pointer-events: none;
 		}
